+++ src/resources/views/layouts/default/private.blade.php
@@ -10,9 +10,9 @@
 
     <title>@yield('title') &middot; {{ $appshell->name }}</title>
 
-    <!-- Styles -->
-    <link href="{{ $appshell->useMix ? mix('/css/appshell.css') : asset('/css/appshell.css') }}" media="all" type="text/css" rel="stylesheet" />
+    @include('appshell::layouts.default._css')
     {!! icon_theme_assets() !!}
+
     @include('appshell::layouts.default._header_includes')
 
     <!-- Scripts -->
@@ -41,7 +41,7 @@
 <!-- Scripts -->
 @stack('scripts')
 
-<script src="{{ $appshell->useMix ? mix('/js/appshell.js') : asset('/js/appshell.js') }}"></script>
+@include('appshell::layouts.default._js')
 
 @include('appshell::layouts.default._scripts')
 @include('appshell::layouts.default._footer_includes')